Return a Future instead of a void
Hello, I tried your package and it is very useful. However, it would be great to be able to await until the modal is closed. It isn't not difficult to do: the Navigator.push already returns a Future, so you just need to return that. :)
Have a good day,
Martin