package:flutter/painting.dart' is unnecessary because all of the used elements are also provided by the import of'package:flutter/cupertino.dart
The name 'DatePickerTheme' is defined in the libraries 'package:flutter/src/material/date_picker_theme.dart (via package:flutter/material.dart)' and 'package:flutter_datetime_picker/src/datetime_picker_theme.dart'.

Using a colon as the separator before a default value is deprecated and will not be supported in language version 3.0 and later.

In '_renderTitleActionsView' function delete itemCount because it is not used